Moved splitters to utils.splitter, added note regarding regex splitter
