+++ tests/splitter_test.py
@@ -7,7 +7,7 @@ import pytest
 
 from wurzel.datacontract import MarkdownDataContract
 from wurzel.steps.splitter import SimpleSplitterStep
-from wurzel.utils.semantic_splitter import (
+from wurzel.utils.splitters.semantic_splitter import (
     SemanticSplitter,
 )
