+++ wurzel/utils/splitters/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2025 Deutsche Telekom AG (opensource@telekom.de)
+#
+# SPDX-License-Identifier: Apache-2.0
