+++ wurzel/utils/splitters/semantic_splitter.py
@@ -16,8 +16,8 @@ from mistletoe.token import Token
 
 from wurzel.datacontract import MarkdownDataContract
 from wurzel.exceptions import MarkdownException
-from wurzel.utils.markdown_table_splitter import MarkdownTableSplitterUtil
-from wurzel.utils.sentence_splitter import SentenceSplitter
+from wurzel.utils.splitters.markdown_table_splitter import MarkdownTableSplitterUtil
+from wurzel.utils.splitters.sentence_splitter import SentenceSplitter
 from wurzel.utils.to_markdown.html2md import MD_RENDER_LOCK
 
 LEVEL_MAPPING = {
